% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.tmleplus}
\alias{print.tmleplus}
\title{Print an object of class \code{tmleplus}}
\usage{
\method{print}{tmleplus}(x)
}
\arguments{
\item{x}{an object of class \code{tmleplus}}
}
\description{
Print an object of class \code{tmleplus}
}
