% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtmle.R
\name{surv_contrast}
\alias{surv_contrast}
\title{Contrast one survtmle object against another (for multilevel exposure)}
\usage{
surv_contrast(tmle_fit_1, tmle_fit_0)
}
\arguments{
\item{tmle_fit_1}{survtmle fit from the exposure of interest}

\item{tmle_fit_0}{survtmle fit from the reference exposure}
}
\value{
A data frame with a row for the difference, standard error, p-value and 95% CI
}
\description{
Contrast one survtmle object against another (for multilevel exposure)
}
