% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtmle.R
\name{surv_eff_mod}
\alias{surv_eff_mod}
\title{Calculate effect modification from survtmle fit objects}
\usage{
surv_eff_mod(tmle_fit_1, tmle_fit_0)
}
\arguments{
\item{tmle_fit_1}{survtmle fit object for only patients with the effect modifier}

\item{tmle_fit_0}{survtmle fit object for only patients without the effect modifier}
}
\value{

}
\description{
This calculations an effect modifier's difference and corresponding standard error for the marginal cumulative incidence using objects obtained from the `survtmle` package.
}
\examples{
# Modified version of the survtmle vignette example
set.seed(1234)
n <- 1000
t_0 <- 6
trt <- rbinom(n, 1, 0.5)
eff <- rbinom(n, 1, 0.5)
adjustVars <- data.frame(W1 = round(runif(n)), W2 = round(runif(n, 0, 2)))
ftime <- round(1 + runif(n, 1, 4) - trt*eff + adjustVars$W1 + adjustVars$W2)
ftype <- round(runif(n, 0, 1))
dat <- data.frame(trt, eff, adjustVars, ftime, ftype)

# Stratify data and run survtmle for all subjects without the effect modifier
dat_noeff <- dat[dat$eff == 0,]
trt <- dat_noeff$trt
adjustVars <- dat_noeff[c("W1", "W2")]
ftime <- dat_noeff$ftime
ftype <- dat_noeff$ftype
fit_noeff <- survtmle::survtmle(ftime = ftime, ftype = ftype,
                  trt = trt, adjustVars = adjustVars,
                  SL.trt = c("SL.glm", "SL.mean", "SL.step"),
                  SL.ftime = c("SL.glm", "SL.mean", "SL.step"),
                  SL.ctime = c("SL.glm", "SL.mean", "SL.step"),
                  method = "hazard", t0 = t_0)

# Run survtmle for all subjects with the effect modifier
dat_eff <- dat[dat$eff == 1,]
trt <- dat_eff$trt
adjustVars <- dat_eff[c("W1", "W2")]
ftime <- dat_eff$ftime
ftype <- dat_eff$ftype
fit_eff <- survtmle::survtmle(ftime = ftime, ftype = ftype,
                  trt = trt, adjustVars = adjustVars,
                  SL.trt = c("SL.glm", "SL.mean", "SL.step"),
                  SL.ftime = c("SL.glm", "SL.mean", "SL.step"),
                  SL.ctime = c("SL.glm", "SL.mean", "SL.step"),
                  method = "hazard", t0 = t_0)

# specify the two tmle fits as arguments
surv_eff_mod(tmle_fit_0 = fit_noeff, tmle_fit_1 = fit_eff)
}
