% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_effect_mod.R
\name{surv_effect_mod}
\alias{surv_effect_mod}
\title{Calculate effect modification from survtmle fit objects}
\usage{
surv_effect_mod(dat_full, mod_var, tmle_fit_0, tmle_fit_1)
}
\arguments{
\item{dat_full}{The full data set containing all observations used to obtain each `tmle_fit_1` and `tmle_fit_0`.}

\item{mod_var}{String containing column name of the effect modifying variable in `dat_full`.}

\item{tmle_fit_0}{survtmle fit object for only patients without the effect modifier}

\item{tmle_fit_1}{survtmle fit object for only patients with the effect modifier}
}
\description{
This calculations an effect modifier's difference and corresponding standard error for the cumulative hazard incidence using objects obtained from the `survtmle` package.
}
