% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survtmle.R
\name{surv_tidy}
\alias{surv_tidy}
\title{Tidy one survtmle object}
\usage{
surv_tidy(tmle_fit)
}
\arguments{
\item{tmle_fit}{A survtmle fit object}
}
\value{
a data frame with the difference in marginal cumulative incidence for a treatment of interest, with standard errors, p-values, and 95% confidence interval
}
\description{
Tidy one survtmle object
}
