% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_ltmle.R
\name{tidy_tmle.ltmleEffectMeasures}
\alias{tidy_tmle.ltmleEffectMeasures}
\title{Extract estimates from a(n) ltmle fit}
\usage{
\method{tidy_tmle}{ltmleEffectMeasures}(fit)
}
\arguments{
\item{fit}{an object of class ltmleEffectMeasures}
}
\value{
a tibble or dataframe of class tmleplus
}
\description{
Extract estimates from a(n) ltmle fit
}
\examples{
rexpit <- function(x) rbinom(n=length(x), size = 1, prob = plogis(x))
n <- 1000
W1 <- rnorm(n)
W2 <- rbinom(n, size=1, prob=0.3)
W3 <- rnorm(n)
A <- rexpit(-1 + 2 * W1 + W3)
Y <- rexpit(-0.5 + 2 * W1^2 + 0.5 * W2 - 0.5 * A + 0.2 * W3 * A - 1.1 * W3)
data <- data.frame(W1, W2, W3, A, Y)
fit <- ltmle(data, Anodes="A", Lnodes=NULL, Ynodes="Y", abar=list(1, 0), SL.library="glm")
tidy_tmle(fit)
}
