% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_tmle.R
\name{tidy_tmle.tmle}
\alias{tidy_tmle.tmle}
\title{Title}
\usage{
\method{tidy_tmle}{tmle}(fit, a = NULL, y = NULL)
}
\arguments{
\item{obs_y}{}
}
\value{

}
\description{
Title
}
\examples{
set.seed(1)
n <- 250
W <- matrix(rnorm(n*3), ncol=3)
A <- rbinom(n,1, 1/(1+exp(-(.2*W[,1] - .1*W[,2] + .4*W[,3]))))
Y <- A + 2*W[,1] + W[,3] + W[,2]^2 + rnorm(n)
tmle_fit <- tmle::tmle(Y,A,W, Q.SL.library = "SL.glm", g.SL.library = "SL.glm")
tidy_tmle(tmle_fit, A, Y)
}
